refactor: rename user to userfeed
