+++ feature/user/src/production/java/com/gcaguilar/randomuser/feature/userfeed/di/NetworkModule.kt
@@ -1,0 +1,31 @@
+package com.gcaguilar.randomuser.feature.userfeed.di
+
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.okhttp.OkHttp
+import io.ktor.client.plugins.HttpTimeout
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.ANDROID
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.serialization.kotlinx.json.json
+import org.koin.dsl.module
+
+val networkModule = module {
+    single<HttpClient> {
+        HttpClient(OkHttp) {
+            install(ContentNegotiation) {
+                json()
+            }
+            install(HttpTimeout) {
+                requestTimeoutMillis = 60000
+                connectTimeoutMillis = 60000
+                socketTimeoutMillis = 60000
+            }
+            install(Logging) {
+                logger = Logger.ANDROID
+                level = LogLevel.ALL
+            }
+        }
+    }
+}
